Update readme
Stripped out unneeded information and links to original creators website ( as well as his name and what seems to be a Copyright of the readme itself. ) Having just rewritten this readme that copyright is not needed. Will retain ( or add if missing ) a copyright to the main source file. It's just ugly having it in the readme. Also I don't care any more about the old Debian install system since it is now history, so not linking to it or mentioning it.